Fix bad type defs creeping into package
Del depends on globby which I found was importing bad definitely typed
files. In particular @types/glob would end up with a weird version of
@types/minimatch.

Bumping del to v6 fixes this. There's also nothing in this bump that
is breaking changes and is the last version before Del is switched
over to esm only (which I personally think is a better move but
noticed prs vetoing it for that reason).